requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

google-genai==1.66.0
instructor>=1.7.0
torch==2.4.0+cpu
pydantic>=2.9.2
streamlit>=1.42.0
python-dotenv
sentence-transformers
scikit-learn
graphviz
pandas
numpy
jsonref>=1.1.0
pm4py
